// src/components/views/settings_contactos/types.ts
export interface ContactoType {
	id?: number;
	nombre: string;
	telefono: string;
	email: string;
	mensaje?: string;
	ip?: string;
	dispositivo?: string;
	useragent_browser?: string;
	useragent_version?: string;
	useragent_os?: string;
	useragent_platform?: string;
	useragent_source?: string;
	date_created?: Date;
}

export interface ContactoTableType extends ContactoType {
	options?: React.ReactNode;
	datecreatedFormated?: string;
}